Move credit card and paypal tests to the apropriate file
