+++ test/Mapper/RequestMapperUTest.php
@@ -34,13 +34,9 @@ namespace WirecardTest\PaymentSdk\Mapper;
 
 use Wirecard\PaymentSdk\Config\Config;
 use Wirecard\PaymentSdk\Config\PaymentMethodConfig;
-use Wirecard\PaymentSdk\Entity\Money;
-use Wirecard\PaymentSdk\Entity\Redirect;
 use Wirecard\PaymentSdk\Mapper\RequestMapper;
 use Wirecard\PaymentSdk\Transaction\CreditCardTransaction;
 use Wirecard\PaymentSdk\Transaction\Operation;
-use Wirecard\PaymentSdk\Transaction\PayPalTransaction;
-use Wirecard\PaymentSdk\Transaction\ThreeDCreditCardTransaction;
 use Wirecard\PaymentSdk\Transaction\Transaction;
 
 class RequestMapperUTest extends \PHPUnit_Framework_TestCase
@@ -57,396 +53,6 @@ class RequestMapperUTest extends \PHPUnit_Framework_TestCase
     protected function setUp()
     {
         $this->mapper = $this->createRequestMapper();
-    }
-
-    public function testPayPalTransaction()
-    {
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'paypal']]],
-            'requested-amount' => ['currency' => 'EUR', 'value' => 24],
-            'notifications' => ['notification' => [['url' => self::EXAMPLE_URL]]],
-            'transaction-type' => 'debit',
-            'cancel-redirect-url' => 'http://www.example.com/cancel',
-            'success-redirect-url' => 'http://www.example.com/success',
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-
-        $redirect = new Redirect('http://www.example.com/success', 'http://www.example.com/cancel');
-
-        $payPalTransaction = new PayPalTransaction();
-        $payPalTransaction->setNotificationUrl(self::EXAMPLE_URL);
-        $payPalTransaction->setRedirect($redirect);
-        $payPalTransaction->setAmount(new Money(24, 'EUR'));
-        $payPalTransaction->setOperation(Operation::PAY);
-        $result = $this->mapper->map($payPalTransaction);
-
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    public function testSslCreditCardTransactionWithTokenId()
-    {
-        $_SERVER['REMOTE_ADDR'] = 'test IP';
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'requested-amount' => ['currency' => 'EUR', 'value' => 24],
-            'ip-address' => 'test IP',
-            'transaction-type' => 'authorization',
-            'card-token' => [
-                'token-id' => '21'
-            ],
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-
-        $cardData = new CreditCardTransaction();
-        $cardData->setTokenId('21');
-        $cardData->setAmount(new Money(24, 'EUR'));
-        $cardData->setOperation(Operation::RESERVE);
-
-        $result = $this->mapper->map($cardData);
-
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    public function testSslCreditCardTransactionWithParentTransactionId()
-    {
-        $_SERVER['REMOTE_ADDR'] = 'test IP';
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'requested-amount' => ['currency' => 'EUR', 'value' => 24],
-            'parent-transaction-id' => 'parent5',
-            'ip-address' => 'test IP',
-            'transaction-type' => 'referenced-authorization',
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-
-        $transaction = new CreditCardTransaction();
-        $transaction->setAmount(new Money(24, 'EUR'));
-        $transaction->setParentTransactionId('parent5');
-        $transaction->setParentTransactionType(Transaction::TYPE_AUTHORIZATION);
-        $transaction->setOperation(Operation::RESERVE);
-        $result = $this->mapper->map($transaction);
-
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    /**
-     * @expectedException \Wirecard\PaymentSdk\Exception\MandatoryFieldMissingException
-     */
-    public function testSslCreditCardTransactionWithoutTokenIdAndParentTransactionId()
-    {
-        $_SERVER['REMOTE_ADDR'] = 'test IP';
-
-        $transaction = new CreditCardTransaction();
-        $transaction->setAmount(new Money(24, 'EUR'));
-        $transaction->setOperation(Operation::RESERVE);
-        $this->mapper->map($transaction);
-    }
-
-    public function testSslCreditCardTransactionWithBothTokenIdAndParentTransactionId()
-    {
-        $_SERVER['REMOTE_ADDR'] = 'test IP';
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'requested-amount' => ['currency' => 'EUR', 'value' => 24],
-            'parent-transaction-id' => 'parent5',
-            'ip-address' => 'test IP',
-            'transaction-type' => 'referenced-authorization',
-            'card-token' => [
-                'token-id' => '33'
-            ],
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-
-        $cardData = new CreditCardTransaction();
-        $cardData->setTokenId('33');
-        $cardData->setAmount(new Money(24, 'EUR'));
-        $cardData->setParentTransactionId('parent5');
-        $cardData->setParentTransactionType(Transaction::TYPE_AUTHORIZATION);
-        $cardData->setOperation(Operation::RESERVE);
-        $result = $this->mapper->map($cardData);
-
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    public function threeDProvider()
-    {
-        return [
-            [
-                Operation::CANCEL,
-                Transaction::TYPE_AUTHORIZATION,
-                Transaction::TYPE_VOID_AUTHORIZATION
-            ],
-            [
-                Operation::RESERVE,
-                null,
-                ThreeDCreditCardTransaction::TYPE_CHECK_ENROLLMENT
-            ],
-            [
-                Operation::RESERVE,
-                ThreeDCreditCardTransaction::TYPE_CHECK_ENROLLMENT,
-                Transaction::TYPE_AUTHORIZATION
-            ],
-            [
-                Operation::RESERVE,
-                Transaction::TYPE_AUTHORIZATION,
-                Transaction::TYPE_REFERENCED_AUTHORIZATION
-            ],
-        ];
-    }
-
-    /**
-     * @param $operation
-     * @param $parentTransactionType
-     * @param $expectedType
-     * @dataProvider threeDProvider
-     */
-    public function testThreeDCreditCardTransaction($operation, $parentTransactionType, $expectedType)
-    {
-        $_SERVER['REMOTE_ADDR'] = 'test IP';
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'requested-amount' => ['currency' => 'EUR', 'value' => 24],
-            'parent-transaction-id' => 'parent54',
-            'ip-address' => 'test IP',
-            'transaction-type' => $expectedType,
-            'card-token' => [
-                'token-id' => '21'
-            ],
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-
-        $money = new Money(24, 'EUR');
-        $creditCardTransaction = new ThreeDCreditCardTransaction();
-        $creditCardTransaction->setTokenId('21');
-        $creditCardTransaction->setTermUrl('https://example.com/r');
-        $creditCardTransaction->setAmount($money);
-        $creditCardTransaction->setParentTransactionId('parent54');
-        $creditCardTransaction->setParentTransactionType($parentTransactionType);
-        $creditCardTransaction->setOperation($operation);
-        $result = $this->mapper->map($creditCardTransaction);
-
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    /**
-     * @expectedException \Wirecard\PaymentSdk\Exception\UnsupportedOperationException
-     */
-    public function testThreeDCreditCardTransactionThrowsUnsupportedOperationException()
-    {
-        $_SERVER['REMOTE_ADDR'] = 'test IP';
-
-        $money = new Money(24, 'EUR');
-        $creditCardTransaction = new ThreeDCreditCardTransaction();
-        $creditCardTransaction->setTokenId('21');
-        $creditCardTransaction->setTermUrl('https://example.com/r');
-        $creditCardTransaction->setAmount($money);
-        $creditCardTransaction->setParentTransactionId('parent54');
-        $creditCardTransaction->setOperation('test');
-        $this->mapper->map($creditCardTransaction);
-    }
-
-    /**
-     * @return array
-     */
-    public function testCancelProvider()
-    {
-        return [
-            [
-                Transaction::TYPE_AUTHORIZATION,
-                Transaction::TYPE_VOID_AUTHORIZATION
-            ],
-            [
-                Transaction::TYPE_REFERENCED_AUTHORIZATION,
-                Transaction::TYPE_VOID_AUTHORIZATION
-            ],
-            [
-                'refund-capture',
-                'void-refund-capture'
-            ],
-            [
-                'refund-purchase',
-                'void-refund-purchase'
-            ],
-            [
-                Transaction::TYPE_CREDIT,
-                'void-credit'
-            ],
-            [
-                CreditCardTransaction::TYPE_PURCHASE,
-                'void-purchase'
-            ],
-            [
-                CreditCardTransaction::TYPE_REFERENCED_PURCHASE,
-                'void-purchase'
-            ],
-            [
-                Transaction::TYPE_CAPTURE_AUTHORIZATION,
-                'void-capture'
-            ]
-        ];
-    }
-
-    /**
-     * @dataProvider testCancelProvider
-     * @param $transactionType
-     * @param $cancelType
-     */
-    public function testCancel($transactionType, $cancelType)
-    {
-        $followupTransaction = new CreditCardTransaction();
-        $followupTransaction->setParentTransactionId('642');
-        $followupTransaction->setParentTransactionType($transactionType);
-        $followupTransaction->setOperation(Operation::CANCEL);
-        $_SERVER['REMOTE_ADDR'] = 'test';
-
-        $result = $this->mapper->map($followupTransaction);
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'parent-transaction-id' => '642',
-            'ip-address' => 'test',
-            'transaction-type' => $cancelType,
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    public function testPayProvider()
-    {
-        return [
-            [
-                Transaction::TYPE_AUTHORIZATION,
-                Transaction::TYPE_CAPTURE_AUTHORIZATION
-            ],
-            [
-                CreditCardTransaction::TYPE_PURCHASE,
-                CreditCardTransaction::TYPE_REFERENCED_PURCHASE
-            ],
-            [
-                null,
-                CreditCardTransaction::TYPE_PURCHASE
-            ]
-        ];
-    }
-
-    /**
-     * @dataProvider testPayProvider
-     * @param $transactionType
-     * @param $payType
-     */
-    public function testPay($transactionType, $payType)
-    {
-        $followupTransaction = new CreditCardTransaction();
-        $followupTransaction->setParentTransactionId('642');
-        $followupTransaction->setParentTransactionType($transactionType);
-        $followupTransaction->setOperation(Operation::PAY);
-        $_SERVER['REMOTE_ADDR'] = 'test';
-
-        $result = $this->mapper->map($followupTransaction);
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'parent-transaction-id' => '642',
-            'ip-address' => 'test',
-            'transaction-type' => $payType,
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    public function testPay3dProvider()
-    {
-        return [
-            [
-                Transaction::TYPE_AUTHORIZATION,
-                Transaction::TYPE_CAPTURE_AUTHORIZATION
-            ],
-            [
-                CreditCardTransaction::TYPE_PURCHASE,
-                CreditCardTransaction::TYPE_REFERENCED_PURCHASE
-            ],
-            [
-                ThreeDCreditCardTransaction::TYPE_CHECK_ENROLLMENT,
-                CreditCardTransaction::TYPE_PURCHASE
-            ],
-            [
-                null,
-                ThreeDCreditCardTransaction::TYPE_CHECK_ENROLLMENT
-            ]
-        ];
-    }
-
-    /**
-     * @dataProvider testPay3dProvider
-     * @param $transactionType
-     * @param $payType
-     */
-    public function testPay3d($transactionType, $payType)
-    {
-        $followupTransaction = new ThreeDCreditCardTransaction();
-        $followupTransaction->setParentTransactionId('642');
-        $followupTransaction->setParentTransactionType($transactionType);
-        $followupTransaction->setOperation(Operation::PAY);
-        $_SERVER['REMOTE_ADDR'] = 'test';
-
-        $result = $this->mapper->map($followupTransaction);
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'parent-transaction-id' => '642',
-            'ip-address' => 'test',
-            'transaction-type' => $payType,
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-        $this->assertEquals(json_encode($expectedResult), $result);
-    }
-
-    /**
-     * @expectedException \Wirecard\PaymentSdk\Exception\MandatoryFieldMissingException
-     */
-    public function testCancelInvalidParentTransaction()
-    {
-        $followupTransaction = new CreditCardTransaction();
-        $followupTransaction->setParentTransactionId('642');
-        $followupTransaction->setParentTransactionType('test');
-        $followupTransaction->setOperation(Operation::CANCEL);
-        $_SERVER['REMOTE_ADDR'] = 'test';
-
-        $this->mapper->map($followupTransaction);
-    }
-
-    public function testCredit()
-    {
-        $followupTransaction = new CreditCardTransaction();
-        $followupTransaction->setParentTransactionId('642');
-        $followupTransaction->setParentTransactionType(Transaction::TYPE_CREDIT);
-        $followupTransaction->setOperation(Operation::CREDIT);
-
-        $_SERVER['REMOTE_ADDR'] = 'test';
-
-        $result = $this->mapper->map($followupTransaction);
-
-        $expectedResult = ['payment' => [
-            'request-id' => '5B-dummy-id',
-            'payment-methods' => ['payment-method' => [['name' => 'creditcard']]],
-            'parent-transaction-id' => '642',
-            'ip-address' => 'test',
-            'transaction-type' => 'credit',
-            'merchant-account-id' => ['value' => 'B612'],
-        ]];
-        $this->assertEquals(json_encode($expectedResult), $result);
     }
 
     public function testMappingWithPaymentMethodSpecificProperties()
